+++ src/app/app.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 
+import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
@@ -14,6 +15,8 @@ import { DebugXmlPopoverComponent } from './search-form/debug-xml-popover/debug-
 import { InputXmlPopoverComponent } from './search-form/input-xml-popover/input-xml-popover.component';
 import { EmbedSearchPopoverComponent } from './search-form/embed-search-popover/embed-search-popover.component';
 import { EmbedStationBoardPopoverComponent } from './station-board/search/embed-station-board-popover/embed-station-board-popover.component';
+import { CustomStopEventXMLPopoverComponent } from './station-board/search/custom-stop-event-xml-popover/custom-stop-event-xml-popover.component';
+import { CustomTripInfoXMLPopoverComponent } from './trip-info/search/custom-trip-info-xml-popover/custom-trip-info-xml-popover.component';
 
 import { JourneyPointInputComponent } from './search-form/journey-point-input/journey-point-input.component';
 import { JourneyResultsComponent } from './journey/journey-results/journey-results.component';
@@ -22,8 +25,7 @@ import { JourneySearchComponent } from './journey/journey-search/journey-search.
 import { MapComponent } from './map/map.component';
 import { ResultTripLegComponent } from './journey/journey-result-row/result-trip-leg/result-trip-leg.component';
 import { SearchFormComponent } from './search-form/search-form.component';
-
-import { CustomStopEventXMLPopoverComponent } from './station-board/search/custom-stop-event-xml-popover/custom-stop-event-xml-popover.component';
+import { TripInfoResultPopoverComponent } from './journey/journey-result-row/result-trip-leg/trip-info-result-popover/trip-info-result-popover.component';
 
 import { StationBoardComponent } from './station-board/station-board.component';
 import { StationBoardInputComponent } from './station-board/input/station-board-input.component';
@@ -31,7 +33,14 @@ import { StationBoardMapComponent } from './station-board/map/station-board-map.
 import { StationBoardResultComponent } from './station-board/result/station-board-result.component';
 import { StationBoardSearchComponent } from './station-board/search/station-board-search.component';
 
+import { TripInfoComponent } from './trip-info/trip-info.component';
+import { TripInfoMapComponent } from './trip-info/map/trip-info-map.component';
+import { TripInfoSearchComponent } from './trip-info/search/trip-info-search.component';
+import { TripInfoResultComponent } from './trip-info/result/trip-info-result.component';
+
 import { TripModeTypeComponent } from './search-form/trip-mode-type/trip-mode-type.component';
+
+import { ServiceStopsComponent } from './shared/components/service-stops.component';
 
 import { EmbedSearchComponent } from './embed/embed-search.component';
 import { EmbedStationBoardComponent } from './embed/embed-station-board.component';
@@ -50,6 +59,7 @@ import { WebHeaderComponent } from './shared/components/web-header';
     ResultTripLegComponent,
     SearchFormComponent,
     TripModeTypeComponent,
+    TripInfoResultPopoverComponent,
 
     StationBoardComponent,
     StationBoardInputComponent,
@@ -57,11 +67,19 @@ import { WebHeaderComponent } from './shared/components/web-header';
     StationBoardResultComponent,
     StationBoardSearchComponent,
 
+    TripInfoComponent,
+    TripInfoMapComponent,
+    TripInfoSearchComponent,
+    TripInfoResultComponent,
+
     DebugXmlPopoverComponent,
     InputXmlPopoverComponent,
     CustomStopEventXMLPopoverComponent,
+    CustomTripInfoXMLPopoverComponent,
     EmbedSearchPopoverComponent,
     EmbedStationBoardPopoverComponent,
+
+    ServiceStopsComponent,
     
     WebFooterComponent,
     WebHeaderComponent,
@@ -69,6 +87,8 @@ import { WebHeaderComponent } from './shared/components/web-header';
     EmbedStationBoardComponent,
   ],
   imports: [
+    CommonModule,
+
     BrowserModule,
     BrowserAnimationsModule,
     HttpClientModule,