+++ src/app/trip-info/trip-info.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'trip-info',
+  templateUrl: './trip-info.component.html',
+  styleUrls: ['./trip-info.component.scss'],
+})
+export class TripInfoComponent implements OnInit {
+  constructor() {
+
+  }
+
+  ngOnInit(): void {
+
+  }
+}