+++ CHANGELOG.md
@@ -5,6 +5,7 @@ OJP-Demo URL: https://opentdatach.github.io/ojp-demo-app/
 ----
 
 26.August 2024
+- display service trip info in a popover / standalone tab - [PR #154](https://github.com/openTdataCH/ojp-demo-app-src/pull/154)
 - paginate results, display prev / next connections - [PR #153](https://github.com/openTdataCH/ojp-demo-app-src/pull/153) 
 - display heading text - [Display Service DestinationText #146](https://github.com/openTdataCH/ojp-demo-app-src/issues/146), [PR #153](https://github.com/openTdataCH/ojp-demo-app-src/pull/153) 
 - use `0.9.36` version of [OJP JS](https://github.com/openTdataCH/ojp-js) SDK